workers/tlp-generator/requirements.txt
sympy>=1.13.0
google-cloud-firestore>=2.14.0
google-cloud-pubsub>=2.19.0
google-cloud-kms>=2.21.0
